+++ actix-tls/src/lib.rs
@@ -4,9 +4,6 @@
 #![doc(html_logo_url = "https://actix.rs/img/logo.png")]
 #![doc(html_favicon_url = "https://actix.rs/favicon.ico")]
 
-#[cfg(feature = "openssl")]
-extern crate tls_openssl as openssl;
-
 #[cfg(feature = "accept")]
 pub mod accept;
 #[cfg(feature = "connect")]